# public/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Fleet Finance Manger",
        "description": "Fleet Finance Manger - API for Android and IOS Development",
        "termsOfService": "All rights reserved to Greefi Technologies",
        "contact": {
            "email": "hello@greefitech.com"
        },
        "version": "1.0.0"
    },
    "host": "localhost:9000",
    "basePath": "/",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/api/staff/create": {
            "post": {
                "tags": [
                    "Staff"
                ],
                "description": "Create New Staff",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Enter Name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "mobile1",
                        "in": "query",
                        "description": "Enter Mobile 1",
                        "required": true,
                        "type": "string",
                        "format": "number"
                    },
                    {
                        "name": "mobile2",
                        "in": "query",
                        "description": "Enter Mobile 2",
                        "type": "string",
                        "format": "number"
                    },
                    {
                        "name": "address",
                        "in": "query",
                        "description": "Enter Address",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "licenceNumber",
                        "in": "query",
                        "description": "Enter licence Number",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "licenceRenewal",
                        "in": "query",
                        "description": "Enter licence Renewal Date [yyyy-mm-dd]",
                        "type": "string"
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Enter Type [manager,driver,cleaner]",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Invalid Data"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/staffs": {
            "get": {
                "tags": [
                    "Staff"
                ],
                "description": "Return all Staffs",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Invalid Data"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/staff/{id}/edit": {
            "get": {
                "tags": [
                    "Staff"
                ],
                "description": "Return particular Staff",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Invalid Data"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/staff/{id}/update": {
            "post": {
                "tags": [
                    "Staff"
                ],
                "description": "Update New Staff",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Enter Name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "mobile1",
                        "in": "query",
                        "description": "Enter Mobile 1",
                        "required": true,
                        "type": "string",
                        "format": "number"
                    },
                    {
                        "name": "mobile2",
                        "in": "query",
                        "description": "Enter Mobile 2",
                        "type": "string",
                        "format": "number"
                    },
                    {
                        "name": "address",
                        "in": "query",
                        "description": "Enter Address",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "licenceNumber",
                        "in": "query",
                        "description": "Enter licence Number",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "licenceRenewal",
                        "in": "query",
                        "description": "Enter licence Renewal Date [yyyy-mm-dd]",
                        "type": "string"
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Enter Type [manager,driver,cleaner]",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Invalid Data"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/staff/{id}/delete": {
            "delete": {
                "tags": [
                    "Staff"
                ],
                "description": "Delete Staff Data",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Invalid Data"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
...
}